合并原作者更新

+++ code/default/launcher/post_update.py
@@ -24,17 +24,18 @@ def run(last_run_version):
         import autorun
         autorun.enable()
 
-    shutil.rmtree(os.path.join(top_path, 'launcher')) # launcher is for auto-update from 2.X
+    if os.path.isdir(os.path.join(top_path, 'launcher')):
+        shutil.rmtree(os.path.join(top_path, 'launcher')) # launcher is for auto-update from 2.X
 
     if older_or_equal(last_run_version, '3.0.4'):
-        xlog.info("migrating to 3.0.5+")
+        xlog.info("migrating to 3.x.x")
         for filename in os.listdir(top_path):
             filepath = os.path.join(top_path, filename)
             if os.path.isfile(filepath):
-                if sys.platform != 'win32' and filename == 'xxnet':
+                if sys.platform != 'win32' and filename == 'start':
                     st = os.stat(filepath)
                     os.chmod(filepath, st.st_mode | stat.S_IEXEC)
-                if not filename.startswith('.') and filename not in ['README.md', 'xxnet', 'xxnet.bat', 'xxnet.vbs']:
+                if not filename.startswith('.') and filename not in ['README.md', 'start', 'start.bat', 'start.vbs']:
                     os.remove(filepath)
             else:
                 if not filename.startswith('.') and filename not in ['code', 'data']:
